Add Input docs
