+++ docs/docs/APIdocs/Layers/Containers/Graph.md
@@ -2,7 +2,7 @@
 
 **Scala:**
 ```scala
-val graph = Graph[Float](Array(Node), Array(Node))
+val graph = Graph(Array(Node), Array(Node))
 ```
 **Python:**
 ```python
@@ -34,17 +34,21 @@ model = Model([Node], [Node])
 
 **Scala example:**
 ```scala
+import com.intel.analytics.bigdl.nn._
+import com.intel.analytics.bigdl.tensor.Tensor
+import com.intel.analytics.bigdl.utils.T
+import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric.NumericFloat
 
-val input1 = Input[Float]()
-val input2 = Input[Float]()
-val cadd = CAddTable[Float]().apply(input1, input2)
-val graph = Graph[Float](Array(input1, input2), cadd)
+val input1 = Input()
+val input2 = Input()
+val cadd = CAddTable().inputs(input1, input2)
+val graph = Graph(Array(input1, input2), cadd)
 
-val output = graph.forward(T(Tensor[Float](T(0.1f, 0.2f, -0.3f, -0.4f)),
-                             Tensor[Float](T(0.5f, 0.4f, -0.2f, -0.1f))))
-val gradInput = graph.backward(T(Tensor[Float](T(0.1f, 0.2f, -0.3f, -0.4f)),
-                                 Tensor[Float](T(0.5f, 0.4f, -0.2f, -0.1f))),
-							   Tensor[Float](T(0.1f, 0.2f, 0.3f, 0.4f)))
+val output = graph.forward(T(Tensor(T(0.1f, 0.2f, -0.3f, -0.4f)),
+                             Tensor(T(0.5f, 0.4f, -0.2f, -0.1f))))
+val gradInput = graph.backward(T(Tensor(T(0.1f, 0.2f, -0.3f, -0.4f)),
+                                 Tensor(T(0.5f, 0.4f, -0.2f, -0.1f))),
+							   Tensor(T(0.1f, 0.2f, 0.3f, 0.4f)))
 
 > println(output)
 output: com.intel.analytics.bigdl.nn.abstractnn.Activity =
@@ -75,24 +79,22 @@ gradInput: com.intel.analytics.bigdl.nn.abstractnn.Activity =
 
 **Python example:**
 ```python
+import numpy as np
 
-fc1 = Linear(4, 2)()
-fc2 = Linear(4, 2)()
-cadd = CAddTable()([fc1, fc2])
-output1 = ReLU()(cadd)
-output2 = Threshold(10.0)(cadd)
-model = Model([fc1, fc2], [output1, output2])
-fc1.element().set_weights([np.ones((4, 2)), np.ones((2, ))])
-fc2.element().set_weights([np.ones((4, 2)) * 2, np.ones((2, )) * 2])
+input1 = Input()
+input2 = Input()
+cadd = CAddTable()(input1, input2)
+model = Model([input1, input2], [cadd])
 
-output = model.forward([np.array([0.1, 0.2, -0.3, -0.4]),
-                        np.array([0.5, 0.4, -0.2, -0.1])])
+output = model.forward([
+    np.ndarray([0.1, 0.2, -0.3, -0.4]),
+    np.ndarray([0.5, 0.4, -0.2, -0.1])])
 
-gradInput = model.backward([np.array([0.1, 0.2, -0.3, -0.4]),
-                            np.array([0.5, 0.4, -0.2, -0.1])],
-                           [np.array([1.0, 2.0]),
-                                    np.array([3.0, 4.0])])
-weights = fc1.element().get_weights()[0]
+gradInput = graph.backward([
+    np.ndarray([0.1, 0.2, -0.3, -0.4]),
+    np.ndarray([0.5, 0.4, -0.2, -0.1])],
+    np.ndarray([0.1, 0.2, 0.3, 0.4])
+)
 
 > output
 [array([ 3.79999971,  3.79999971], dtype=float32),
