+++ pyspark/bigdl/nn/layer.py
@@ -1630,6 +1630,27 @@ class CosineDistance(Layer):
                  bigdl_type="float"):
         super(CosineDistance, self).__init__(None, bigdl_type)
 
+class Input(Layer):
+
+    '''
+    Input layer do nothing to the input tensors, just pass them. It should be used as input node
+    when the first layer of your module accepts multiple tensors as inputs.
+
+    Each input node of the graph container should accept one tensor as input. If you want a module
+    accepting multiple tensors as input, you should add some Input module before it and connect
+    the outputs of the Input nodes to it.
+
+    Please note that the return is not a layer but a Node containing input layer.
+
+    >>> input = Input()
+    creating: createCosineDistance
+    '''
+
+    def __init__(self,
+                 bigdl_type="float"):
+        layer = super(Input, self).__init__(None, bigdl_type)
+        layer()
+
 
 class DotProduct(Layer):
