+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/python/api/PythonBigDL.scala
@@ -268,6 +268,10 @@ class PythonBigDL[T: ClassTag](implicit ev: TensorNumeric[T]) extends Serializab
     Echo[T]()
   }
 
+  def createInput(): Input[T] = {
+    new Input[T]()
+  }
+
   def createLogSoftMax(): LogSoftMax[T] = {
     LogSoftMax[T]()
   }
